refactor: ♻️ move config mixins into package
